changes to sign care recognition and central control to stop at stoip sign.
