+++ dependencies.txt
@@ -147,10 +147,10 @@ netifaces==0.10.9
 nose==1.3.7
 notebook==6.4.8
 notebook-shim==0.1.0
-numpy==1.19.2
+numpy==1.24.4
 open3d==0.15.1
 opencv-contrib-python==4.5.3.56
-opencv-python==4.4.0.44
+opencv-python==4.8.1.78
 packaging==20.4
 pandas==1.4.1
 pandocfilters==1.5.0
@@ -313,15 +313,15 @@ tk==0.1.0
 toml==0.10.1
 tomlkit==0.10.0
 topic-tools==1.15.9
-torch==1.10.2
-torchvision==0.11.3
+torch==2.1.1
+torchvision==0.16.1
 tornado==4.5.3
-tqdm==4.62.3
+tqdm==4.66.1
 traitlets==5.1.1
 Twisted==20.3.0
 txaio==21.2.1
 typed-ast==1.4.1
-typing-extensions==3.7.4.3
+typing-extensions==4.3.0
 unique_id==1.0.6
 urdfdom-py==0.4.5
 urllib3==1.26.4
