+++ src/central_control/full_system.launch
@@ -11,8 +11,8 @@
       This is needed for automated testing. -->
   <node name="central_control" pkg="central_control" type="central_control.py" />
   <node name="lane_keep_assist" pkg="lane_keep_assist" type="send_lane.py" />
-  <node name="car_detect" pkg="sign_car_recognition" type="sign_detect.py" />
-  <node name="stop_sign_detect" pkg="sign_car_recognition" type="new_sign_detect.py" />
+  <node name="object_detector" pkg="sign_car_recognition" type="object_detection.py" />
+  <node name="stop_sign_detector" pkg="sign_car_recognition" type="stop_sign_detection.py" />
   <node name="object_detect" pkg="lidar_object_recognition_2" type="new_lidar_object_detect.py"/>
 
   <node name="cruise_control" pkg="cruise_control" type="cruise_control.py" />
